+++ src/shoutit/api/v3/serializers/shout.py
@@ -13,19 +13,20 @@ from rest_framework.reverse import reverse
 from common.constants import POST_TYPE_REQUEST, POST_TYPE_OFFER
 from shoutit.controllers import shout_controller
 from shoutit.models import Shout, Currency, InactiveShout
-from shoutit.utils import upload_image_to_s3, debug_logger
+from shoutit.utils import upload_image_to_s3, debug_logger, blank_to_none
 from .base import LocationSerializer, VideoSerializer
 from .profile import ProfileSerializer
 from .tag import CategorySerializer
 from ..exceptions import ERROR_REASON
+empty_input = {'allow_blank': True, 'allow_null': True, 'required': False}
 
 
 class ShoutSerializer(serializers.ModelSerializer):
     type = serializers.ChoiceField(source='get_type_display', choices=['offer', 'request'], help_text="*")
     location = LocationSerializer(
         help_text="Defaults to user's saved location, Passing the `latitude` and `longitude` is enough to calculate new location properties")
-    title = serializers.CharField(source='item.name', min_length=4, max_length=50, default='', allow_blank=True, allow_null=True, help_text="Max 50 characters")
-    text = serializers.CharField(min_length=10, max_length=1000, default='', allow_null=True, allow_blank=True, help_text="Max 1000 characters")
+    title = serializers.CharField(source='item.name', min_length=4, max_length=50, help_text="Max 50 characters", **empty_input)
+    text = serializers.CharField(min_length=10, max_length=1000, help_text="Max 1000 characters", **empty_input)
     price = serializers.IntegerField(source='item.price', allow_null=True, required=False, help_text="Value in cents")
     available_count = serializers.IntegerField(default=1, help_text="Only used for Offers")
     is_sold = serializers.BooleanField(default=False, help_text="Only used for Offers")
@@ -96,7 +97,9 @@ class ShoutSerializer(serializers.ModelSerializer):
     def to_representation(self, instance):
         if instance.muted or instance.is_disabled:
             return InactiveShout().to_dict
-        return super(ShoutSerializer, self).to_representation(instance)
+        ret = super(ShoutSerializer, self).to_representation(instance)
+        blank_to_none(ret, ['title', 'text'])
+        return ret
 
 
 class ShoutDetailSerializer(ShoutSerializer):
@@ -106,7 +109,7 @@ class ShoutDetailSerializer(ShoutSerializer):
     reply_url = serializers.SerializerMethodField(
         help_text="URL to reply to this shout if possible, not set for shout owner")
     conversations = serializers.SerializerMethodField()
-    mobile = serializers.CharField(min_length=4, max_length=20, allow_blank=True, default='', write_only=True)
+    mobile = serializers.CharField(min_length=4, max_length=20, write_only=True, **empty_input)
     mobile_hint = serializers.CharField(read_only=True)
     is_mobile_set = serializers.BooleanField(read_only=True)