+++ src/shoutit/models/item.py
@@ -3,6 +3,7 @@ from django.contrib.postgres.fields import ArrayField
 from django.db import models
 from common.constants import ITEM_STATE_AVAILABLE
 from shoutit.models.base import UUIDModel
+from shoutit.utils import none_to_blank
 
 
 class Item(UUIDModel):
@@ -20,10 +21,7 @@ class Item(UUIDModel):
         return unicode(self.name[:30]) if self.name else '[Item]'
 
     def clean(self):
-        if self.name is None:
-            self.name = ""
-        if self.description is None:
-            self.description = ""
+        none_to_blank(self, ['name', 'description'])
 
     @property
     def thumbnail(self):